feat: Add component for exporting data to Sheets
Introduces a new React component, `ExportToSheets`, responsible for generating and exporting data in a spreadsheet-friendly format. This component includes utility functions for formatting currency and percentages, as well as a helper for converting strings to colors. It also defines a `DRETableHTML` component to render a Profit and Loss statement (DRE) table with conditional styling.